Fixes buggy context menu behavior
PopupMenu behaved inproperly in some cases. It would cause the window to
lock up. It has been substituted with Swing's JPopupMenu. To get around
the fact that Swing uses lightweight components, and the canvas is a
Heavyweight component, currPopup is forced into Heavyweight mode. This
ensures that the context menu is always drawn above the canvas. Custom
checks had to be made to close the context menu when you click out of
it, and attaching it to the parent container of canvas ensures
compatibility.